Filesystem Cache for Map Tiles
- server stores tile images directly on file system
- this will enable easy access without hitting database cache
- enables more seamless offline access to cached tiles
- minimizes requests to external tile server
