+++ .npmignore
@@ -7,3 +7,4 @@ logs
 Dockerfile
 docker-compose.yml
 triton-compose.yml
+tiles
