+++ makefile
@@ -20,6 +20,8 @@ pack:
 		-o web/public/platform/a.js
 
 	browserify platform/vendor/display.js \
+		platform/vendor/storage.js \
+		platform/modules/network/database.js \
 		platform/modules/application/director.js \
 		platform/modules/application/app.js \
 		platform/modules/application/view.js \
@@ -29,10 +31,8 @@ pack:
 	browserify platform/helpers/array.js \
 		platform/helpers/string.js \
 		platform/helpers/math.js \
-		platform/vendor/storage.js \
 		platform/modules/network/item.js \
 		platform/modules/atlas/marker.js \
-		platform/modules/network/database.js \
 		platform/modules/user/feed.js \
 		platform/modules/user/user.js \
 		platform/modules/organization/organization.js \
