+++ package.json
@@ -14,11 +14,11 @@
         "gun": "^0.9.999995",
         "gun-tag": "*",
         "helmet": "*",
+        "node-fetch": "*",
         "latlon-geohash": "*",
         "leaflet": ">=1.3.4",
         "leaflet.locatecontrol": "*",
         "pouchdb-browser": "^7.0.0",
-        "request": "*",
         "shortid": "*",
         "simple-node-logger": "*",
         "typeface-montserrat": "0.0.54",
